Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -101,9 +101,13 @@ router.put(
     imagePath
   });
 
-  Post.updateOne({ _id: req.params.id }, post)
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
     .then(result => {
-      res.status(200).json({message: 'Update successful'});
+      if (result.nModified > 0) {
+        res.status(200).json({ message: 'Update successful!' })
+      } else {
+        res.status(401).json({ message: 'Not Authorized!' })
+      }
     })
     .catch(err => console.log(err));
 });
@@ -123,12 +127,13 @@ router.get('/:id', (req, res, next) => {
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id})
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId})
     .then(result => {
-      console.log(result);
-      res.status(200).json({
-        message: 'Post deleted!'
-      });
+      if (result.nModified > 0) {
+        res.status(200).json({ message: 'Deletion successful!' })
+      } else {
+        res.status(401).json({ message: 'Not Authorized!' })
+      }
     });
 });
